# connectors/unique_search_proxy/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Unique Search Proxy",
    "description": "Unified web egress proxy for search engines and crawlers. Use **Try it out** on `/v1/search` and `/v1/crawl` \u2014 pick an example from the request-body dropdown (snippets-only Google search, crawl with HTML markdown, etc.). Requires provider env vars (e.g. `GOOGLE_SEARCH_API_KEY`, `GOOGLE_SEARCH_ENGINE_ID`) for live calls.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/v1/configuration/search-engines/{engine_id}/json-schema": {
      "get": {
        "tags": [
          "configuration"
        ],
        "summary": "JSON Schema for one search-engine config",
        "operationId": "search_engine_config_schema_v1_configuration_search_engines__engine_id__json_schema_get",
        "parameters": [
          {
            "name": "engine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Engine Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderJsonSchemaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/configuration/search-engines/{engine_id}/default-config": {
      "get": {
        "tags": [
          "configuration"
        ],
        "summary": "Default deployment config for one search engine",
        "operationId": "search_engine_default_config_v1_configuration_search_engines__engine_id__default_config_get",
        "parameters": [
          {
            "name": "engine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Engine Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderDefaultConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/configuration/search-engines/{engine_id}/call-schema": {
      "post": {
        "tags": [
          "configuration"
        ],
        "summary": "JSON Schema for POST /v1/search from deployment config",
        "operationId": "search_engine_call_schema_v1_configuration_search_engines__engine_id__call_schema_post",
        "parameters": [
          {
            "name": "engine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Engine Id"
            }
          },
          {
            "name": "strict",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "When true (default), the call schema marks every exposed field as required with no defaults (for strict LLM JSON generation). When false, optional provider knobs keep nullable types and defaults.",
              "default": true,
              "title": "Strict"
            },
            "description": "When true (default), the call schema marks every exposed field as required with no defaults (for strict LLM JSON generation). When false, optional provider knobs keep nullable types and defaults."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleConfig"
              },
              "examples": {
                "default_projection": {
                  "summary": "Call schema (query + fetchSize only)",
                  "description": "Default LLM projection when optional params are not exposed.",
                  "value": {
                    "engine": "google",
                    "fetchSize": 10,
                    "safe": "active"
                  }
                },
                "with_exposed_params": {
                  "summary": "Call schema (+ gl)",
                  "description": "`gl.expose=true` adds `gl` to the LLM call schema; `dateRestrict` stays merge-only.",
                  "value": {
                    "engine": "google",
                    "fetchSize": 10,
                    "safe": "active",
                    "gl": {
                      "expose": true,
                      "value": "ch"
                    },
                    "dateRestrict": {
                      "expose": false,
                      "value": "d7"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchCallSchemaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/configuration/crawlers/{crawler_id}/json-schema": {
      "get": {
        "tags": [
          "configuration"
        ],
        "summary": "JSON Schema for one crawler config",
        "operationId": "crawler_config_schema_v1_configuration_crawlers__crawler_id__json_schema_get",
        "parameters": [
          {
            "name": "crawler_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Crawler Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderJsonSchemaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/configuration/crawlers/{crawler_id}/default-config": {
      "get": {
        "tags": [
          "configuration"
        ],
        "summary": "Default deployment config for one crawler",
        "operationId": "crawler_default_config_v1_configuration_crawlers__crawler_id__default_config_get",
        "parameters": [
          {
            "name": "crawler_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Crawler Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderDefaultConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExposableEI": {
        "properties": {
          "expose": {
            "type": "boolean",
            "title": "Expose",
            "description": "When true, this parameter is included on the LLM call JSON Schema."
          },
          "value": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "e",
                  "i"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Value",
            "description": "Admin default merged into each search when not ``None``."
          }
        },
        "type": "object",
        "required": [
          "expose",
          "value"
        ],
        "title": "ExposableEI"
      },
      "ExposableStr": {
        "properties": {
          "expose": {
            "type": "boolean",
            "title": "Expose",
            "description": "When true, this parameter is included on the LLM call JSON Schema."
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value",
            "description": "Admin default merged into each search when not ``None``."
          }
        },
        "type": "object",
        "required": [
          "expose",
          "value"
        ],
        "title": "ExposableStr"
      },
      "GoogleConfig": {
        "properties": {
          "engine": {
            "type": "string",
            "const": "google",
            "title": "Search engine",
            "description": "Provider discriminator; must be `google` for this config.",
            "default": "google"
          },
          "fetchSize": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Fetchsize",
            "description": "Default result count merged into each search request",
            "default": 10
          },
          "timeout": {
            "type": "integer",
            "maximum": 600.0,
            "minimum": 1.0,
            "title": "Timeout",
            "description": "Request timeout in seconds",
            "default": 30
          },
          "searchEngineId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search engine ID (cx)",
            "description": "Google Programmable Search Engine ID (`cx`). Defaults from `GOOGLE_SEARCH_ENGINE_ID` when deployed. Not sent as a query parameter (resolved at runtime)."
          },
          "safe": {
            "type": "string",
            "enum": [
              "active",
              "off"
            ],
            "title": "Safe search",
            "description": "SafeSearch level for every search: `active` (default) or `off`. Applied on all requests unless the call body overrides it.",
            "default": "active"
          },
          "gl": {
            "$ref": "#/components/schemas/ExposableStr",
            "title": "Geolocation (gl)",
            "description": "Two-letter ISO 3166-1 alpha-2 country code (Google `gl`). Set `value` for a fixed default; set `expose` so the LLM may override per query."
          },
          "hl": {
            "$ref": "#/components/schemas/ExposableStr",
            "title": "Interface language (hl)",
            "description": "Language for snippets/UI (Google `hl`). `value` + `expose` behave like `gl`."
          },
          "lr": {
            "$ref": "#/components/schemas/ExposableStr",
            "title": "Language restrict (lr)",
            "description": "Document language restrict (Google `lr`), e.g. `lang_en`. `value` + `expose` behave like `gl`."
          },
          "dateRestrict": {
            "$ref": "#/components/schemas/ExposableStr",
            "title": "Date restrict",
            "description": "Google `dateRestrict` recency filter (`d7`, `m1`, \u2026). `value` + `expose` behave like `gl`."
          },
          "exactTerms": {
            "$ref": "#/components/schemas/ExposableStr",
            "title": "Exact terms",
            "description": "Phrase every hit must contain (Google `exactTerms`). `value` + `expose` behave like `gl`."
          },
          "excludeTerms": {
            "$ref": "#/components/schemas/ExposableStr",
            "title": "Exclude terms",
            "description": "Phrase that must not appear (Google `excludeTerms`). `value` + `expose` behave like `gl`."
          },
          "fileType": {
            "$ref": "#/components/schemas/ExposableStr",
            "title": "File type",
            "description": "File extension filter (Google `fileType`), e.g. `pdf`. `value` + `expose` behave like `gl`."
          },
          "siteSearch": {
            "$ref": "#/components/schemas/ExposableStr",
            "title": "Site search",
            "description": "Site or domain (Google `siteSearch`). Pair with `siteSearchFilter`. `value` + `expose` behave like `gl`."
          },
          "siteSearchFilter": {
            "$ref": "#/components/schemas/ExposableEI",
            "title": "Site search filter",
            "description": "With `siteSearch`: `i` = include only that site, `e` = exclude. `value` + `expose` behave like `gl`."
          },
          "sort": {
            "$ref": "#/components/schemas/ExposableStr",
            "title": "Sort",
            "description": "Sort expression (Google `sort`), e.g. `date`. `value` + `expose` behave like `gl`."
          }
        },
        "type": "object",
        "title": "GoogleConfig",
        "description": "Single source of truth for Google deployment + derived request/LLM surfaces."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProviderDefaultConfigResponse": {
        "properties": {
          "providerId": {
            "type": "string",
            "title": "Providerid"
          },
          "defaultConfig": {
            "additionalProperties": true,
            "type": "object",
            "title": "Defaultconfig",
            "description": "Default deployment config (camelCase keys)"
          }
        },
        "type": "object",
        "required": [
          "providerId",
          "defaultConfig"
        ],
        "title": "ProviderDefaultConfigResponse"
      },
      "ProviderJsonSchemaResponse": {
        "properties": {
          "jsonSchema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Jsonschema",
            "description": "JSON Schema for provider deployment configuration"
          },
          "providerId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Providerid",
            "description": "Set when the schema is for a single provider"
          }
        },
        "type": "object",
        "required": [
          "jsonSchema"
        ],
        "title": "ProviderJsonSchemaResponse"
      },
...
      "SearchCallSchemaResponse": {
        "properties": {
          "engine": {
            "type": "string",
            "title": "Engine",
            "description": "Search engine id"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "description": "Engine mode (e.g. standard) for observability and tooling"
          },
          "snippetOnly": {
            "type": "boolean",
            "title": "Snippetonly",
            "description": "When true, search hits are snippet-only; use POST /v1/crawl for bodies"
          },
          "callSchema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Callschema",
            "description": "JSON Schema for the engine call model on POST /v1/search"
          }
        },
        "type": "object",
        "required": [
          "engine",
          "mode",
          "snippetOnly",
          "callSchema"
        ],
        "title": "SearchCallSchemaResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}